[None][chore] Split sampler package into common + strategy layers
Introduce sampler_common.py as the shared base layer (generic helpers,
per-request queries reading LlmRequest/SamplingConfig, request->strategy
bridge) and rename sampling_utils.py to sampler_strategy.py (the strategy
types, resolution and dispatch it already held). The request-query helpers
(_unwrap_singleton, int_tensor, add_token, _get_*_beam_width,
_request_get_sampling_params, _request_strategy, UtilsSamplingParams) move
out of sampler.py into these layers.

Import graph stays acyclic and single-directional:
sampler -> sampler_common -> sampler_strategy -> ops. Update the package
__init__ lazy-forward submodule set accordingly. No behavior change.
